Revert "Feature/code line node"
