+++ packages/lexical-clipboard/src/clipboard.ts
@@ -28,13 +28,13 @@ import {
   $createParagraphNode,
   $getRoot,
   $getSelection,
-  $isCodeLineNode,
   $isDecoratorNode,
   $isElementNode,
   $isGridCellNode,
   $isGridNode,
   $isGridRowNode,
   $isGridSelection,
+  $isLineBreakNode,
   $isRangeSelection,
   $isTextNode,
   $parseSerializedNode,
@@ -184,7 +184,7 @@ function $basicInsertStrategy(
       ($isDecoratorNode(node) && !node.isTopLevel()) ||
       ($isElementNode(node) && node.isInline()) ||
       $isTextNode(node) ||
-      $isCodeLineNode(node)
+      $isLineBreakNode(node)
     ) {
       if (currentBlock === null) {
         currentBlock = $createParagraphNode();
