+++ packages/lexical-playground/src/plugins/CodeHighlightPlugin.ts
@@ -14,7 +14,7 @@ export default function CodeHighlightPlugin(): JSX.Element | null {
   const [editor] = useLexicalComposerContext();
 
   useEffect(() => {
-    return registerCodeHighlighting(editor, 10000);
+    return registerCodeHighlighting(editor);
   }, [editor]);
 
   return null;
