+++ packages/lexical/flow/Lexical.js.flow
@@ -639,27 +639,6 @@ declare export function $isLineBreakNode(
 ): boolean %checks(node instanceof LineBreakNode);
 
 /**
- * LexicalCodeLineNode
- */
-
-declare export class CodeLineNode extends LexicalNode {
-  static getType(): string;
-  static clone(node: CodeLineNode): CodeLineNode;
-  constructor(key?: NodeKey): void;
-  getTextContent(): '\n';
-  createDOM(): HTMLElement;
-  updateDOM(): false;
-  static importJSON(
-    serializedCodeLineNode: SerializedCodeLineNode,
-  ): CodeLineNode;
-  exportJSON(): SerializedLexicalNode;
-}
-declare export function $createCodeLineNode(): CodeLineNode;
-declare export function $isCodeLineNode(
-  node: ?LexicalNode,
-): boolean %checks(node instanceof CodeLineNode);
-
-/**
  * LexicalRootNode
  */
 
@@ -819,7 +798,7 @@ declare export function $getRoot(): RootNode;
 declare export function $isLeafNode(
   node: ?LexicalNode,
 ): boolean %checks(node instanceof TextNode ||
-  node instanceof CodeLineNode ||
+  node instanceof LineBreakNode ||
   node instanceof DecoratorNode);
 declare export function $setCompositionKey(
   compositionKey: null | NodeKey,
@@ -888,12 +867,6 @@ export type SerializedParagraphNode = {
 export type SerializedLineBreakNode = {
   ...SerializedLexicalNode,
   type: 'linebreak',
-  ...
-};
-
-export type SerializedCodeLineNode = {
-  ...SerializedLexicalNode,
-  type: 'codeline',
   ...
 };
