+++ packages/lexical/src/__tests__/unit/LexicalEditor.test.tsx
@@ -17,8 +17,8 @@ import {
   TableRowNode,
 } from '@lexical/table';
 import {
-  $createCodeLineNode,
   $createGridSelection,
+  $createLineBreakNode,
   $createNodeSelection,
   $createParagraphNode,
   $createTextNode,
@@ -1744,7 +1744,7 @@ describe('LexicalEditor tests', () => {
     await editor.update(() => {
       const root = $getRoot();
       const child = root.getLastDescendant();
-      child.insertAfter($createCodeLineNode());
+      child.insertAfter($createLineBreakNode());
     });
 
     expect(fn).toHaveBeenCalledTimes(2);
